Learnspot/pyproject.toml
[project]
name = "learnspot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "ipykernel>=7.2.0",
    "openai>=2.32.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]
